fix removing images

+++ app/service/book/BookService.php
@@ -225,8 +225,10 @@ class BookService
 
     public function saveImage(CoverInfoParameters $coverInfoParameters, $book)
     {
-        $this->imageService->removeImage($book->coverImage);
         if($coverInfoParameters->getImage() != null){
+            if($book->image != 'images/questionCover.png'){
+                $this->imageService->removeImage($book->coverImage);
+            }
             if($coverInfoParameters->getImageSaveType() == ImageSaveType::UPLOAD){
                 $book->coverImage = $this->imageService->saveUploadImage($coverInfoParameters->getImage(),
                     $book->title);
